+++ crates/macros/src/config/field.rs
@@ -1,6 +1,6 @@
 use crate::common::{Field, FieldValue};
 use proc_macro2::{Literal, TokenStream};
-use quote::{quote, ToTokens, TokenStreamExt};
+use quote::{ToTokens, TokenStreamExt, quote};
 
 impl Field<'_> {
     pub fn generate_default_value(&self) -> TokenStream {
@@ -29,13 +29,16 @@ impl Field<'_> {
 
         #[cfg(feature = "env")]
         {
-            let value = if let Some(parse_env) = &self.args.parse_env {
-                quote! {
-                    parse_env_value(#env, #parse_env)?
+            let value = match &self.args.parse_env {
+                Some(parse_env) => {
+                    quote! {
+                        parse_env_value(#env, #parse_env)?
+                    }
                 }
-            } else {
-                quote! {
-                    default_env_value(#env)?
+                _ => {
+                    quote! {
+                        default_env_value(#env)?
+                    }
                 }
             };