+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## Unreleased
+
+#### 🚀 Updates
+
+- Updated to Rust v1.85 and 2024 edition.
+
+#### 🐞 Fixes
+
+- Fixed an issue where importing `Result` would break our generated macro code.
+
+#### ⚙️ Internal
+
+- Updated dependencies.
+
 ## 0.17.11
 
 #### 🚀 Updates